updates in agent.py script
modifications in the get_maxQ and choose_action functions, i.e. set the agent to pick a random available action with epsilon probability, and to pick the q-highest available action with 1-epsilon probability. I modified and used get_maxQ so that it will return the maximum Q value for an action given a state.